requirements.txt
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
boto3>=1.26.0
fastapi>=0.95.0
uvicorn>=0.22.0
streamlit>=1.22.0
plotly>=5.14.0
python-dotenv>=1.0.0
mangum>=0.17.0
bottleneck>=1.3.6
backtrader>=1.9.76.123
bt>=0.2.9
matplotlib>=3.7.0
scikit-learn>=1.4.0
seaborn>=0.13.0
pyarrow>=15.0.0
pytest>=8.0.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
scikit-optimize>=0.9.0
tqdm>=4.66.0
joblib>=1.3.0
lightgbm>=4.1.0 